Finished tests and further refined logging
