--- build_and_test.ps1
@@ -1,13 +1,0 @@
-if (Test-Path env:VIRTUAL_ENV)
-{
-    ./build.ps1
-    ./test.ps1
-}
-else{
-    ./venv/scripts/Activate.ps1
-    ./build.ps1
-    ./test.ps1
-    deactivate
-}
-
-
